LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/relaxing-the-same-origin-restriction/document_domain_feature_policy.tentative.sub-expected.txt



PASS Default "document-domain" feature policy ["*"] allows the top-level document.
PASS Default "document-domain" feature policy ["*"] allows same-origin iframes.
PASS Default "document-domain" feature policy ["*"] allows cross-origin iframes.
FAIL Feature policy "document-domain" can be disabled in cross-origin iframes using "allow" attribute. assert_equals: expected "#SecurityError" but got "#OK"
FAIL Feature policy "document-domain" can be disabled in same-origin iframes using "allow" attribute. assert_equals: expected "#SecurityError" but got "#OK"
